Addressing Windows Intel 64bit compile issues (Trilinos #5869 and #5870)
There where problems with ambiguous abs() calls.  The fix is to cast the
argument type 'double', then call.  This has not been tested in the Windows
Intel 64bit build but I guarantee it will fix the build.

Build/Test Cases Summary
Enabled Packages: EpetraExt, Triutils
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=678,notpassed=0 (42.46 min)
1) SERIAL_RELEASE => passed: passed=657,notpassed=0 (20.21 min)

+++ packages/epetraext/src/EpetraExt_MatrixMatrix_mult_A_B.cpp
@@ -78,7 +78,7 @@ static inline int C_estimate_nnz(const Epetra_CrsMatrix & A, const Epetra_CrsMat
   int Aest=(A.NumMyRows()>0)? A.NumMyNonzeros()/A.NumMyRows():100;
   int Best=(B.NumMyRows()>0)? B.NumMyNonzeros()/B.NumMyRows():100;
 
-  int nnzperrow=(int)(sqrt(Aest) + sqrt(Best) - 1);
+  int nnzperrow=(int)(sqrt((double)Aest) + sqrt((double)Best) - 1);
   nnzperrow*=nnzperrow;
  
   return (int)(A.NumMyRows()*nnzperrow*0.75 + 100);
